requirements.txt
dotenv
requests
google-genai
langchain-mistralai

